Convert domain to string when saving user dashboard
This commit fixes an issue where the dashboard would be broken when the user removed a section through the web interface.

The domain would be wrongly interpreted when generating the new view using QWeb.
The solution is to deal with the domain the same way it is dealt with modifiers (attrs).